+++ server/models/recomment.js
@@ -16,15 +16,15 @@ module.exports = (sequelize, DataTypes) => {
     id: {
         primaryKey: true,
         autoIncrement: true,
-        type: DataTypes.SMALLINT,
+        type: DataTypes.INTEGER,
         allowNull: false
     },
     comment_id: {
-        type: DataTypes.SMALLINT,
+        type: DataTypes.INTEGER,
         allowNull: false,
     },
     user_id: {
-        type: DataTypes.SMALLINT,
+        type: DataTypes.INTEGER,
         allowNull: false,
     },
     content: {
@@ -32,7 +32,7 @@ module.exports = (sequelize, DataTypes) => {
         allowNull: false,
     },
     reports_count: {
-        type: DataTypes.SMALLINT,
+        type: DataTypes.INTEGER,
         defaultValue: 0,
         allowNull: false,
     },
@@ -49,11 +49,11 @@ module.exports = (sequelize, DataTypes) => {
         allowNull: false,
     },
     adopted_status: {
-        type: DataTypes.TINYINT(1),
+        type: DataTypes.INTEGER,
         allowNull: false,
     },
     deleted_status: {
-        type: DataTypes.TINYINT(1),
+        type: DataTypes.INTEGER,
         allowNull: false,
     },
   }, {
